+++ frontend/src/components/BackendEndpoint.tsx
@@ -1,31 +1,17 @@
-import React from "react";
-
 function getHostAndPort() {
     const hostname = window?.location?.hostname ?? "localhost";
     const port = window?.location?.port ?? "3001";
-
     return hostname + ":" + port;
 }
 
-function getBackendURL() {
+export function getBackendURL() {
     const protocol = window?.location?.protocol ?? "http";
-
-    let backendURL = protocol + "//" + getHostAndPort();
-    
-    if (process.env.REACT_APP_API_HOST !== undefined) {
-      backendURL = process.env.REACT_APP_API_HOST;
-    }
-
-    return backendURL;
+    // if REACT_APP_API_HOST was set in env, use the defined URL, otherwise derive backend URL from current window.location in browser
+    return process.env.REACT_APP_API_HOST ?? (protocol + "//" + getHostAndPort());
 }
 
-function getWsBackendURL() {
-    let wsBackendURL = "ws://" + getHostAndPort();
-    
-    if (process.env.REACT_APP_WS_HOST !== undefined) {
-      wsBackendURL = process.env.REACT_APP_WS_HOST;
-    }
-
-    return wsBackendURL;
+export function getWsBackendURL() {
+    // if REACT_APP_WS_HOST was set in env, use the defined URL, otherwise derive backend URL from current window.location in browser
+    return process.env.REACT_APP_WS_HOST ?? ("ws://" + getHostAndPort());
 }
